+++ misskey-api/src/endpoint/admin.rs
@@ -31,3 +31,7 @@ pub mod remove_abuse_user_report;
 #[cfg(feature = "12-49-0")]
 #[cfg_attr(docsrs, doc(cfg(feature = "12-49-0")))]
 pub mod resolve_abuse_user_report;
+
+#[cfg(feature = "12-80-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "12-80-0")))]
+pub mod ad;